Remove hard dependency on DropWizard Metrics backend.
* Make DropWizard Metrics an optional metrics aggregation backend
* Make Spring Boot Actuator an optional metrics aggregation backend

This is largely done because Spring Boot 1.3 will be shipping with support
for pushing metrics to Statsd and JMX stores which was most of the reason
for using DropWizard in the first place.

By making each of the backends optional we don't force Spring applications
to use another lib for little benefit. We also don't force non-Spring
applications to pull in the entire world of Spring.

Fixes #2